Fix bugs and add category,transactions,budgets Controller

+++ app/Http/Controllers/BudgetsController.php
@@ -20,13 +20,6 @@ class BudgetsController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'user_id' => 'required|exists:users,id',
-            'category_id' => 'required|exists:categories,id',
-            'amount' => 'required|numeric|min:0',
-            'description' => 'nullable|string|max:255',
-        ]);
-
         $budget = Budgets::create($request->all());
 
         return response()->json([
@@ -38,22 +31,29 @@ class BudgetsController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Budgets $Budget)
+    public function show($id)
     {
-        return response()->json($Budget);
+        $budget = Budgets::find($id);
+
+        if (!$budget) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Category not found'
+            ], 404);
+        }
+
+        return response()->json([
+            'success' => true,
+            'data' => $budget
+        ], 200);
     }
+
 
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, Budgets $budget)
     {
-        $request->validate([
-            'user_id' => 'sometimes|exists:users,id',
-            'category_id' => 'sometimes|exists:categories,id',
-            'amount' => 'sometimes|numeric|min:0',
-            'description' => 'nullable|string|max:255',
-        ]);
 
         $budget->update($request->all());
 
@@ -66,9 +66,22 @@ class BudgetsController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Budgets $budget)
+    public function destroy($id)
     {
+       $budget = Budgets::find($id);
+
+        if (!$budget) {
+            return response()->json([
+                'success' => false,
+                'message' => 'User not found'
+            ], 404);
+        }
+
         $budget->delete();
-        return response()->json(['message'=> 'Budget deleted successfully'],204);
+
+        return response()->json([
+            'success' => true,
+            'message' => 'User deleted successfully'
+        ], 200);
     }
 }
